pyproject.toml
[project]
name = "fnirsi-dps-150"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
]

[project.scripts]
dps150 = "dps150.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = ["pytest>=8"]
